:recycle: refactor: rename outbound package to out
