+++ kok-api/src/test/java/com/kok/kokapi/station/application/service/StationServiceTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.kok.kokapi.common.config.ServiceTestConfig;
 import com.kok.kokapi.common.util.DatabaseCleanerExtension;
-import com.kok.kokapi.station.adapter.outbound.persistence.StationRepository;
+import com.kok.kokapi.station.adapter.out.persistence.StationRepository;
 import com.kok.kokcore.station.domain.entity.Station;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
